application/backend/streaming-agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-ai-slim[bedrock]>=1.70.0
boto3>=1.42.0
httpx>=0.27.0
Pillow>=10.0.0
